requirements.txt
# Core packages
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
joblib>=1.1.0

# Data visualization
seaborn>=0.11.0
plotly>=5.5.0

# Deep learning
tensorflow>=2.8.0

# Technical analysis
ta>=0.10.0  # Technical analysis library for indicators

# Data processing
scipy>=1.7.0
statsmodels>=0.13.0  # For time series analysis

# Utilities
tqdm>=4.62.0
argparse>=1.4.0
PyYAML>=6.0

# Web reporting
jinja2>=3.0.0  # For HTML report templating
websockets>=10.0  # For WebSocket communication

# Sentiment analysis and Reddit API
praw
vaderSentiment